Move middlewares and separated tests
